VITSCAN-main/backend/backend.main/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
google-genai>=1.13.0,<2.0.0
pydantic-settings>=2.4.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
bcrypt>=4.1.3,<5.0.0
email-validator>=2.2.0,<3.0.0
